% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_across_ages.R
\name{aggregate_across_ages}
\alias{aggregate_across_ages}
\title{Aggregate value across age groups for standardized timeseries}
\usage{
aggregate_across_ages(df)
}
\arguments{
\item{df}{(\link[tibble:tibble]{tibble::tibble}) A standardized tibble output from \link{standardize_timeseries}}
}
\value{
A standardized data frame with timeseries of outcome in terms of 100K individuals (aggregated across age groups)
}
\description{
Aggregate value across age groups for standardized timeseries
}
