% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_outcome.R
\name{lookup_outcome}
\alias{lookup_outcome}
\title{Decode outcomes for each model into a standardized name}
\usage{
lookup_outcome()
}
\value{
(tibble) A lookup table of outcomes
}
\description{
Mapping between internal (standardized) outcome label, labels for each generic model, and a human-readible label (e.g., for plotting)
}
\examples{
lookup_outcome()
}
