% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_outcome.R
\name{lookup_outcome_label}
\alias{lookup_outcome_label}
\title{Decode standardized outcome names into plot-ready labels}
\usage{
lookup_outcome_label()
}
\value{
(tibble) A lookup table of outcome labels
}
\description{
Decode standardized outcome names into plot-ready labels
}
\examples{
lookup_outcome_label()
}
