% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_outcomes.R
\name{lookup_outcomes}
\alias{lookup_outcomes}
\title{Lookup table for outcome names}
\usage{
lookup_outcomes()
}
\value{
(tibble) A lookup table for outcomes
}
\description{
Lookup table for outcomes
}
\details{
Mapping between internal (standardized) outcome label, labels for each generic model, and a human-readible label (e.g., for plotting)
}
\examples{
lookup_outcomes()
}
