% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_ci_across_iterations}
\alias{summarize_ci_across_iterations}
\title{Summarize confidence interval of values across iterations}
\usage{
summarize_ci_across_iterations(df, ci)
}
\arguments{
\item{df}{(tibble) Standardized timeseries}

\item{ci}{(numeric) Proportion for confidence interval (e.g., 0.95 corresponds to the 95\% confidence interval)}
}
\value{
\link[tibble:tibble]{tibble::tibble}
}
\description{
Summarize confidence interval of values across iterations
}
