% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_outcomes_by_peak}
\alias{summarize_outcomes_by_peak}
\title{Summarize outcomes by peak}
\usage{
summarize_outcomes_by_peak(df)
}
\arguments{
\item{df}{(tibble) Standardized timeseries}
}
\value{
\link[tibble:tibble]{tibble::tibble}
}
\description{
Summarize outcomes by peak
}
