% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_peak_value}
\alias{summarize_peak_value}
\title{Summarize peak value}
\usage{
summarize_peak_value(df, days_horizon = NULL)
}
\arguments{
\item{df}{(tibble) Standardized timeseries}

\item{days_horizon}{(numeric) Calculate summary after this many days}
}
\value{
\link[tibble:tibble]{tibble::tibble}
}
\description{
Summarize peak value
}
