% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_range_across_iterations}
\alias{summarize_range_across_iterations}
\title{Summarize range of values across iterations}
\usage{
summarize_range_across_iterations(df)
}
\arguments{
\item{df}{(tibble) Standardized timeseries}
}
\value{
\link[tibble:tibble]{tibble::tibble}
}
\description{
Summarize range of values across iterations
}
