% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_total}
\alias{summarize_total}
\title{Summarize total value}
\usage{
summarize_total(df, days_horizon = NULL)
}
\arguments{
\item{df}{(tibble) Standardized timeseries}

\item{days_horizon}{(numeric) Calculate summary after this many days}
}
\value{
\link[tibble:tibble]{tibble::tibble}
}
\description{
Summarize total value
}
