% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_filename.R
\name{parse_filename}
\alias{parse_filename}
\title{Parse output filename}
\usage{
parse_filename(filepath)
}
\arguments{
\item{filepath}{file path (or just the filename)}
}
\value{
A list with fields model ("epac" or "abm"), file type ("timeseries" or "summary-metrics"), and scenario id (last part of filename)
}
\description{
Parse standardized filename from model output. Compatible with filenames of format \verb{[[model]]_[[filetype]]_[[id_scenario]].csv}.
}
\examples{
parse_filename("epac_timeseries_smallpox.csv")
}
