% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_timeseries.R
\name{standardize_timeseries}
\alias{standardize_timeseries}
\title{Standardize timeseries across model outputs}
\usage{
standardize_timeseries(filepath, pop.factor_abm = 20)
}
\arguments{
\item{filepath}{Path to a file}

\item{pop.factor_abm}{(only used if input is an ABM file) Population normalization factor for the ABM to rescale to outcomes per 100K individuals (since ABM can be run for different population sizes)}
}
\value{
A standardized data frame with timeseries of outcome in terms of 100K individuals
}
\description{
Standardize timeseries across model outputs
}
