% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_grouped_value}
\alias{summarize_grouped_value}
\title{Summarize a grouped value}
\usage{
summarize_grouped_value(df, fun, days_horizon)
}
\arguments{
\item{df}{(tibble) Standardized timeseries}

\item{fun}{(function) Summary function}

\item{days_horizon}{(numeric) Calculate summary after this many days}
}
\value{
\link[tibble:tibble]{tibble::tibble}
}
\description{
Summarize a grouped value
}
